requirements.txt
asgiref==3.8.1
asttokens==3.0.0
colorama==0.4.6
decorator==5.2.1
Django==5.2.1
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
drf-yasg==1.21.10
executing==2.2.0
Faker==37.1.0
inflection==0.5.1
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Markdown==3.8
matplotlib-inline==0.1.7
packaging==25.0
parso==0.8.4
prompt_toolkit==3.0.51
pure_eval==0.2.3
Pygments==2.19.1
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
tzdata==2025.2
uritemplate==4.1.1
validate-docbr==1.10.0
wcwidth==0.2.13
